tests/wpt/web-platform-tests/signed-exchange/check-cert-request.tentative.html

<!DOCTYPE html>
<title>Verify that certificate request has correct parameters.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  const sxgUrl = get_host_info().HTTPS_ORIGIN + '/signed-exchange/resources/sxg/check-cert-request.sxg';
  const message = await openSXGInIframeAndWaitForMessage(t, sxgUrl);
  assert_equals(message.location, innerURLOrigin() + '/signed-exchange/resources/inner-url.html');
  assert_false(message.is_fallback);
}, 'Verify that certificate request has correct parameters.');

</script>
</body>
